+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Added the `Marker` class and expose the `onChangeBackMarker`, 
 `onChangeFrontMarker`, `addMarker`, and `getMarkers` members on the 
 `EditSession` class.
+- Added the `Fold`, `FoldLine`, `Placeholder` and `RangeList` classes.
+- Added the `Folding` abstract class and have `EditSession` implement it. 
+- Added the `onChangeFold` event stream to `EditSession`.
 
 ## 0.1.7+1.8.2014